[refactor] Extract `TestHelper+Loader`
